Getting DPI-C export functions to work
This is a bit of a hack.

I found that an error I was seeing when calling velcomp was originating from a
missing main function which should have been included by "tbx_main.cxx". Why
this file is not being included is still a mystery to me, but copying into my
project seems to have done the trick.

Signed-off-by: Stephano Cetola <stephanoc@gmail.com>